backend/requirements.txt
psycopg2-binary
flask-corsflask-sqlalchemy
flask
python-dotenv
flask-jwt-extended
werkzeug
requests
gunicorn

